accurately report available? for compact index
The Fetcher system expects #available? to return false if it is impossible to get meaningful results from the CompactIndex fetcher. Given that constraint, we move the step of updating /versions and parsing out the info checksums into the #available? method. It shouldn’t take any more time, since that is a required step for any use of the compact index, and we can use the FallbackError that results from a 404 to definitively indicate that the CompactIndex is not available.